+++ NEWS.md
@@ -1,5 +1,9 @@
 # performance 0.6.1
 
+## General
+
+* `check_collinearity()` gains a `verbose` argument, to toggle warnings and messages.
+
 ## Bug fixes
 
 * Fixed examples, now using suggested packages only conditionally.